feat(flask_app): 💎 Change code structure
Changed code structure,
files like app.py or test_app.py will be deleted soon so if you have any problems or questions about new code structure, message me
