+++ tests/test_api.py
@@ -2,6 +2,7 @@ import os
 import sys
 import pytest
 
+# Ensure the 'flask_app' directory is in the Python path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../flask_app')))
 
 from api import app, db
